+++ src/core/app/tiler.hpp
@@ -7,9 +7,9 @@ typedef struct
     void (*setApp)(App *app);
     void (*renderApp)(bool forced);  
     Component* (*contains)(signed short x, signed short y);
-    void (*onTouch)(Component *component, signed short x, signed short y, unsigned int tickCount); 
-    void (*onMove)(Component *component, signed short x, signed short y, unsigned int tickCount); 
-    void (*onRelease)(Component *component, signed short x, signed short y, unsigned int tickCount); 
+    void (*onTouch)(Component *component, signed short x, signed short y, unsigned long tickCount); 
+    void (*onMove)(Component *component, signed short x, signed short y, unsigned long tickCount); 
+    void (*onRelease)(Component *component, signed short x, signed short y, unsigned long tickCount); 
     void (*onGesture)(Gesture gesture); 
     void (*onButton)();
 } Tiler;
